+++ tests/Validator.test.js
@@ -1,12 +1,13 @@
 import Validator from 'Validator.js';
+import { VALID_DATES } from './defaults';
 
 let validator = new Validator;
 
 describe('Validator', () => {
 
     test('it can validate a value based on a default rule', () => {
-        let validString = validator.validateDefaultRule('field_name', 'This is a string', 'string');
-        let validObject = validator.validateDefaultRule('field_name', {key: "value"}, 'object');
+        let validString = validator.validateDefaultRule('field_name', 'string', {field_name: 'This is a string'});
+        let validObject = validator.validateDefaultRule('field_name', 'object', {field_name: {key: "value"}});
 
         expect(validString).toBeTruthy();
         expect(validObject).toBeTruthy();
@@ -16,8 +17,8 @@ describe('Validator', () => {
         let rule = (name, value, fail) => {
             return value % 2 == 0 || fail('Failed Rule');
         }
-        let validEven = validator.validateCustomRule('field_name', 24, rule);
-        let invalidEven = validator.validateCustomRule('field_name', 23, rule);
+        let validEven = validator.validateCustomRule('field_name', rule, {field_name: 24});
+        let invalidEven = validator.validateCustomRule('field_name', rule, {field_name: 23});
         
         expect(validEven).toBeTruthy();
         expect(invalidEven).toBe('Failed Rule');
@@ -27,14 +28,14 @@ describe('Validator', () => {
 
     test('it can validate a value based on multiple rules', () => {
         let validationTrue = validator.validate(
-            'field_name',
-            'This is a string with more than 10 characters', 
-            ['string', 'min:10']
+            'field_name', 
+            ['string', 'min:10'],
+            {field_name: 'This is a string with more than 10 characters'}
         );
         let validationFalse = validator.validate(
-            'field_name',
-            'This is a string with more than 10 characters', 
-            ['string', 'max:10']
+            'field_name', 
+            ['string', 'max:10'],
+            {field_name: 'This is a string with more than 10 characters'}
         );
 
         expect(validationTrue.valid).toBeTruthy();
@@ -58,146 +59,156 @@ describe('Validator', () => {
          */
 
         // Array
-        expect(validator.validate('field_name', [1,2], ['array']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['array'], {field_name: [1,2]}).valid).toBeTruthy();
 
         // Between
-        expect(validator.validate('field_name', '2', ['between:1,3']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['between:1,3'], {field_name: '2'}).valid).toBeTruthy();
 
         // Boolean
-        expect(validator.validate('field_name', false, ['boolean']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['boolean'], {field_name: false}).valid).toBeTruthy();
 
         // Confirmed
-        validator.setData({
+        expect(validator.validate('password', ['confirmed'], {
             password: 'secret',
             password_confirmation: 'secret',
-            other_field: 'super_secret'
-        });
-        expect(validator.validate('password', 'secret', ['confirmed']).valid).toBeTruthy();
-        expect(validator.validate('other_field', 'super_secret', ['confirmed']).valid).toBeFalsy();
+        }).valid).toBeTruthy();
+
+        expect(validator.validate('other_field', ['confirmed'], {
+            password: 'secret'
+        }).valid).toBeFalsy();
+
+        // Date
+        expect(validator.validate('field_name', ['date'], {field_name: '12 May 1992'}).valid).toBeTruthy();
+
+        // Date Equals
+        expect(validator.validate('first', ['date_equals:second'], {
+            first:  new Date('12 May 1992'),
+            second:  new Date('12 May 1992')
+        }).valid).toBeTruthy();
 
         // Different
-        validator.setData({
+        expect(validator.validate('first', ['different:second'], {
             first: 'hello',
             second: 'world'
-        });
-        expect(validator.validate('first', 'hello', ['different:second']).valid).toBeTruthy();
-        expect(validator.validate('first', 'world', ['different:second']).valid).toBeFalsy();
+        }).valid).toBeTruthy();
+
+        expect(validator.validate('first', ['different:second'], {
+            first: 'hello',
+            second: 'hello'
+        }).valid).toBeFalsy();
 
         // Email
-        expect(validator.validate('field_name', 'email@example.com', ['email']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['email'], {field_name: 'email@example.com'}).valid).toBeTruthy();
 
         // Equals
-        expect(validator.validate('field_name', 2, ['equal:2']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['equal:2'], {field_name: 2}).valid).toBeTruthy();
+
+        // File
+        let file = new File([''], 'avatar.jpeg', {
+            type: 'image/jpeg',
+            lastModified: 1575297668218
+        });
+
+        expect(validator.validate('avatar', ['file'], {avatar: file}).valid).toBeTruthy();
 
         // Filled
-        validator.setData({
-            email: 'email@example.com',
-        });
-        expect(validator.validate('email', 'email@example.com', ['filled']).valid).toBeTruthy();
+        expect(validator.validate('email', ['filled'], {email: 'email@example.com'}).valid).toBeTruthy();
 
         // Greater Than
-        expect(validator.validate('field_name', 2, ['gt:1']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['gt:1'], {field_name: 2}).valid).toBeTruthy();
 
         // Greater Than Or Equals
-        expect(validator.validate('field_name', 2, ['gte:2']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['gte:2'], {field_name: 2}).valid).toBeTruthy();
 
         // In
-        expect(validator.validate('field_name', 'foo', ['in:foo,bar']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['in:foo,bar'], {field_name: 'foo'}).valid).toBeTruthy();
 
         // Integer
-        expect(validator.validate('field_name', -12, ['integer']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['integer'], {field_name: -12}).valid).toBeTruthy();
 
         // Length
-        expect(validator.validate('field_name', 'test', ['length:4']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['length:4'], {field_name: 'test'}).valid).toBeTruthy();
 
         // Less Than
-        expect(validator.validate('field_name', 2, ['lt:3']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['lt:3'], {field_name: 2}).valid).toBeTruthy();
 
         // Less Than Or Equals
-        expect(validator.validate('field_name', 2, ['lte:2']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['lte:2'], {field_name: 2}).valid).toBeTruthy();
 
         // Max
-        expect(validator.validate('field_name', 4, ['max:5']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['max:5'], {field_name: 4}).valid).toBeTruthy();
 
         // Min
-        expect(validator.validate('field_name', 4, ['min:3']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['min:3'], {field_name: 4}).valid).toBeTruthy();
 
         // Null
-        expect(validator.validate('field_name', null, ['null']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['null'], {field_name: null}).valid).toBeTruthy();
 
         // Numeric
-        expect(validator.validate('field_name', 1, ['numeric']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['numeric'], {field_name: 1}).valid).toBeTruthy();
 
         // Object
-        expect(validator.validate('field_name', {}, ['object']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['object'], {field_name: new Object}).valid).toBeTruthy();
 
         // Required
-        expect(validator.validate('field_name', 'test', ['required']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['required'], {field_name: 'test'}).valid).toBeTruthy();
 
         // Required If
-        validator.setData({
+        expect(validator.validate('company_id', ['required_if:has_company,true'], {
             has_company: true,
             company_id: 1,
-        });
-        expect(validator.validate('company_id', 1, ['required_if:has_company,true']).valid).toBeTruthy();
+        }).valid).toBeTruthy();
 
         // Required Unless
-        validator.setData({
+        expect(validator.validate('company_id', ['required_unless:has_company,true'], {
             has_company: true,
             company_id: 1,
-        });
-        expect(validator.validate('company_id', 1, ['required_unless:has_company,true']).valid).toBeTruthy();
+        }).valid).toBeTruthy();
 
-        validator.setData({
+        expect(validator.validate('company_id', ['required_unless:has_company,false'], {
             has_company: true,
-        });
-        expect(validator.validate('company_id', 1, ['required_unless:has_company,false']).valid).toBeFalsy();
+        }).valid).toBeFalsy();
 
         // Required With
-        validator.setData({
+        expect(validator.validate('cars', ['required_with:has_cars'], {
             has_cars: true,
             cars: ['toyota camry', 'honda civic'],
-        });
-        expect(validator.validate('cars', ['toyota camry', 'honda civic'], ['required_with:has_cars']).valid).toBeTruthy();
+        }).valid).toBeTruthy();
 
-        validator.setData({
+        expect(validator.validate('cars', ['required_with:has_cars'], {
             has_cars: true,
-        });
-        expect(validator.validate('cars', ['toyota camry', 'honda civic'], ['required_with:has_cars']).valid).toBeFalsy();
+        }).valid).toBeFalsy();
 
         // Required With All
-        validator.setData({
+        expect(validator.validate('cars', ['required_with_all:has_cars,loves_cars'], {
             loves_cars: true,
             has_cars: true,
             cars: ['toyota camry', 'honda civic'],
-        });
-        expect(validator.validate('cars', ['toyota camry', 'honda civic'], ['required_with_all:has_cars,loves_cars']).valid).toBeTruthy();
+        }).valid).toBeTruthy();
 
-        validator.setData({
+        expect(validator.validate('cars', ['required_with_all:has_cars,loves_cars'], {
             has_cars: true,
             cars: ['toyota camry', 'honda civic']
-        });
-        expect(validator.validate('cars', ['toyota camry', 'honda civic'], ['required_with_all:has_cars,loves_cars']).valid).toBeFalsy();
+        }).valid).toBeFalsy();
 
-        validator.setData({
+        expect(validator.validate('cars', ['required_with_all:has_cars'], {
             has_cars: true,
-        });
-        expect(validator.validate('cars', ['toyota camry', 'honda civic'], ['required_with_all:has_cars']).valid).toBeFalsy();
+        }).valid).toBeFalsy();
 
         // Same
-        expect(validator.validate('field_name', 2, ['same:2']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['same:2'], {field_name: 2}).valid).toBeTruthy();
 
         // String
-        expect(validator.validate('field_name', 'test', ['string']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['string'], {field_name: 'test'}).valid).toBeTruthy();
 
         // URL
-        expect(validator.validate('field_name', 'https://google.com', ['url']).valid).toBeTruthy();
+        expect(validator.validate('field_name', ['url'], {field_name: 'https://google.com'}).valid).toBeTruthy();
     })
 
-    
 
     test('it can set form data', () => {
 
+        var validator = new Validator;
         validator.setData({
             test: 'field'
         });